### 2023-03-02 #5207 @editorialbot
[PRE REVIEW]: Spikeometric - Linear Non-Linear Cascade Spiking Neural Networks with PyTorch Geometric
**Submitting author:** @lepmik (<a href="http://orcid.org/0000-0002-4262-5549">Mikkel Elle Lepperød</a>)
**Repository:** https://github.com/bioAI-Oslo/Spikeometric
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
