### 2023-03-04 #5207 @arfon
@lepmik – I'm going to put this submission through a scope review to solicit the opinion of the wider editorial team on whether this is in scope for JOSS.
