### 2023-04-20 3e49c7c
Fixed a bug in tuning BernoulliGLM and added test to verify
