### 2023-04-24 #5207 @TomDonoghue
@jbytecode - thanks for the invitation - I think this one might be a bit outside of my area of expertise, as I don't have much hands-on experience with this kind of neural network model, so probably someone else would be a better reviewer for this one.
