### 2023-04-24 #5207 @jbytecode
👋👋👋 Dear @clinssen  👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**Spikeometric - Linear Non-Linear Cascade Spiking Neural Networks with PyTorch Geometric**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5207).

Thank you in advance!
