### 2023-05-26 #5451 @clinssen
Thanks to the authors for the development effort and to the editors of JOSS for their excellent work! Please find my initial review below.

Spikeometric is a Python package that has been in development since at least May 30th, 2022. Its total lines of code count is 667, thereby extending the PyTorch Geometric package, which in turn extends PyTorch. It provides functionality related to several classes of linear/nonlinear models in computational neuroscience. The software appears mature, with API documentation, user documentation, tutorials, CI, issue tracker etc. Although small in scale, this is due to its focus on a relatively simple framework (that of linear/nonlinear models), including several model variants and methods for providing stimulus to the network.

Unfortunately, the documentation does not contain any (comparative) benchmarks. Therefore claims of "fast" cannot be verified; however, this claim is quite generic (perhaps "fast" refers to the low time investment necessary by end users?), and is not essential to the software. There does not seem to be a demonstration/example of a very large, "real world size" network. There is no textual comparison to other (similar) tools in the field.

The stimulus generators seem rather basic and it would be interesting to provide, for example, input from files. The generators also seem to generate all spike times ahead of the simulation rather than generating spikes "live" while the simulation is ongoing.

There are two issues with the DOIs that were already observed by the editorial bot.
