### 2023-06-02 #5451 @lepmik
Hi, a bit unsure of what you mean here. Should we start responding to reviews immediately? - I assumed we should wait until all reviews are done.
