### 2023-06-02 #5451 @jbytecode
@lepmik - Sure, you can start to discuss and apply the changes. The review process of JOSS is interactive so you don't need to wait until the whole stuff ends.
