### 2023-06-16 8acb9fb
Fixed typos in documentation and adjusted to match improved stimulus api
