### 2023-06-16 5830ff1
Added a loaded stimulus class and a new base stimulus class that handles batching of stimulus
