### 2023-06-16 05f6c69
Removed stimulus_masks and models now bump stimulus batches after each simulation
