### 2023-06-16 9114e0a
Added tests for loaded stimulus and batched stimulus
