### 2023-06-25 f83bea4
Made it possible to store network state with different dtypes
