### 2023-06-25 1605130
Made it possible to store network state with different dtypes
