### 2023-06-26 #5451 @JakobSonstebo
@jbytecode I'm almost done addressing the points brought up by @clinssen. It is true that the 'fast' emphasis is more about friendliness than speed, so I will change the wording in the article to make this apparent. However, I'm also adding a benchmarking section in the documentation. Next, I have added an example where spikeometric is used to run simulations of a big network imported from the MICrONS synapse dataset from the Allen Institute. This will be merged into the main branch as soon as possible. For the stimulus, there is now a class supporting loading stimulus from a file, and an improved system for batching stimulus. Finally, I've fixed the DOI issues identified by the bot and am in the process of adding a discussion of other similar tools to the article. I hope to have everything done by the end of the week at the latest.
