### 2023-06-27 a63ef27
Added section on other work to paper and fixed DOI issue
