+++ paper.bib
@@ -27,7 +27,7 @@
 	title = {Inferring causal connectivity from pairwise recordings and optogenetics},
 	elocation-id = {463758},
 	year = {2020},
-	doi = {8.1101/463760},
+	doi = {10.1101/463760},
 	publisher = {Cold Spring Harbor Laboratory},
 	abstract = {To study how the brain works mechanistically, neuroscientists want to quantify causal interactions between neurons, typically requiring perturbations. When using optogenetic interventions, multiple neurons are usually perturbed which produces a confound {\textendash} any of the stimulated neurons can have affected the postsynaptic neuron making it challenging to discern which of the neurons produced the causal effect. Here we show how such confounds produce large biases, and we explain how they can be reduced by combining instrumental variable (IV) and difference in differences (DiD) techniques from econometrics. The interaction between stimulation and the absolute refractory period of the neuron produces a weak, approximately random signal which can be exploited to estimate causal transmission probability. On simulated neural networks, we find that estimates using ideas from IV and DiD outperform na{\"\i}ve techniques suggesting that methods from causal inference can be useful when studying neural interactions.Competing Interest StatementThe authors have declared no competing interest.},
 	URL = {https://www.biorxiv.org/content/early/2020/03/07/463760},
@@ -108,7 +108,106 @@ url={https://doi.org/10.1038/nature07140}
   journal={Network: Computation in Neural Systems},
   volume={15},
   number={4},
+  doi={10.1088/0954-898x_15_4_002},
   pages={243--262},
   year={2004},
   publisher={Taylor \& Francis}
+}
+
+@ARTICLE{Gewaltig:NEST,
+  author  = {Marc-Oliver Gewaltig and Markus Diesmann},
+  title   = {NEST (NEural Simulation Tool)},
+  journal = {Scholarpedia},
+  year    = {2007},
+  volume  = {2},
+  pages   = {1430},
+  number  = {4}
+}
+
+@book{carnevale_hines_2006,
+  place={Cambridge},
+  title={The NEURON Book},
+  DOI={10.1017/CBO9780511541612},
+  publisher={Cambridge University Press},
+  author={Carnevale, Nicholas T. and Hines, Michael L.},
+  year={2006}
+}
+
+@article {10.7554/eLife.47314,
+  article_type = {journal},
+  title = {Brian 2, an intuitive and efficient neural simulator},
+  author = {Stimberg, Marcel and Brette, Romain and Goodman, Dan FM},
+  editor = {Skinner, Frances K and Calabrese, Ronald L and Skinner, Frances K and Zeldenrust, Fleur and Gerkin, Richard C},
+  volume = 8,
+  year = 2019,
+  month = {aug},
+  pub_date = {2019-08-20},
+  pages = {e47314},
+  citation = {eLife 2019;8:e47314},
+  doi = {10.7554/eLife.47314},
+  url = {https://doi.org/10.7554/eLife.47314},
[…]
