+++ paper.md
@@ -30,18 +30,19 @@ bibliography: paper.bib
 
 To understand the dynamics of the brain, computational neuroscientists often study smaller scale networks using simple cascade point-process models such as the Linear-Nonlinear-Poisson (LNP) model and the Generalized Linear Model (GLM) [@paninski2004maximum; @gerstner_kistler_naud_paninski_2014; @10.3389/fnsys.2016.00109].
 
-Stochastic models can give key insights into the behavior of a network on a systems level without explicitly modelling the subcellular mechanisms of each neuron. They lack some biological plausibility on the neuron level but have been shown to enjoy nice convexity properties, which can be fitted to observed spike data [@paninski2004maximum]. Traditionally, these are used as encoding models. For example, to study how multi-neuron systems process incoming stimuli [@Pillow2008]. Recently, these models have been used as generative models for inverse problems mapping activity to connectivity. As an example, [@Das2020] assessed bias and reconstruction errors in this setting.
+Stochastic models can give key insights into the behavior of a network on a systems level without explicitly modeling the subcellular mechanisms of each neuron. They lack some biological plausibility on the neuron level but have been shown to enjoy nice convexity properties, which can be fitted to observed spike data [@paninski2004maximum]. Traditionally, these are used as encoding models. For example, to study how multi-neuron systems process incoming stimuli [@Pillow2008]. Recently, these models have been used as generative models for inverse problems mapping activity to connectivity. As an example, [@Das2020] assessed bias and reconstruction errors in this setting.
 
-This software expands on a simulator developed for testing novel reconstruction techniques using methods from the causal inference literature [@Lepperod463758]. It provides tunable generative models in a flexible framework based on PyTorch. The primary use case, so far, has been as a data-generator for inverse problems, but the framework can easily accommodate more complicated models for encoding applications.
+This software expands on a simulator developed for testing novel reconstruction techniques using methods from the causal inference literature [@Lepperod463758]. It provides tunable generative models in a flexible framework based on PyTorch. The primary use case, so far, has been as a data-generator for inverse problems, but the framework can easily accommodate more complicated models for encoding applications. 
 
 # Statement of need
 
-Linear non-linear cascade models are much used in computational neuroscience and come in many flavors. Typical examples are SRM, LNP, GLM [@Gerstner2008; @gerstner_kistler_naud_paninski_2014; @10.3389/fnsys.2016.00109]. What unites these models is that they all model the spike response of a network through the same cascade-like sequence of steps. At each time step, a neuron receives input from its environment and converts it to a firing rate by a nonlinear function. This firing rate parametrizes a probability distribution from which spikes are drawn. This contrasts hard-threshold-based models, in which spikes are emitted whenever a variable representing the membrane potential exceeds a certain value.
-Although these models share many of the same underlying principles, no unifying framework currently permits easy implementation and direct comparison. Moreover a fast and convenient data-generation tool is currently lacking.
+Linear non-linear cascade models are much used in computational neuroscience and come in many flavors. Typical examples are SRM, LNP, GLM [@Gerstner2008; @gerstner_kistler_naud_paninski_2014; @10.3389/fnsys.2016.00109]. What unites these models is that they all model the spike response of a network through the same cascade-like sequence of steps. At each time step, a neuron receives input from its environment and converts it to a firing rate by a nonlinear function. This firing rate parametrizes a probability distribution from which spikes are drawn. This contrasts hard-threshold-based models, in which spikes are emitted whenever a variable representing the membrane potential exceeds a certain value. Although these models share many of the same underlying principles, no unifying framework currently permits easy implementation and direct comparison. 
+
[…]
