### 2023-06-27 8e78f20
Create draft-pdf.yml which builds paper on push to paper branch
