### 2023-06-28 2e762d3
Stores state as dtype consistently throughout simulation
