### 2023-07-01 2d03f90
Fixed unsafe behavior listing all files in directory
