### 2023-07-02 #5451 @JakobSonstebo
Hey again. I've released a new version of spikeometric (1.0.2) on PyPI with the new features. The paper is also updated and in the paper branch. The only thing missing is a benchmarking section of the documentation, which is taking a while because of limited GPU availability. It is, however, in the works. What are the next steps?
