### 2023-08-31 bioAI-Oslo/Spikeometric#12 @Saran-nns
Dependency mismatch between pip package and source
Hi team,

As part of Joss [Review](https://github.com/openjournals/joss-reviews/issues/5451)

Installation through `pip` as mentioned in the documentation works, but when i try to install it from source `pip install -r requirements.txt`, there are version conflicts, showing there are mismatch in required dependencies between the python package and source.

`ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
spikeometric 1.0.2 requires seaborn<0.13.0,>=0.12.2, but you have seaborn 0.11.2 which is incompatible.
Successfully installed seaborn-0.11.2`

Also, when i install the package using the source in `colab` environment, received several dependencies conflict with the latest stable versions in colab. See https://colab.research.google.com/drive/1l8Y3So5uBXl4kA_6jvpkvyZOILVjdrHK#scrollTo=7S0xGBPE_pqK

Looks like package requirements/dependencies are not defined clearly or at least not at their latest stable versions.

thanks
saran
