+++ requirements.txt
@@ -1,138 +1,20 @@
-alabaster==0.7.13 ; python_version >= "3.8" and python_version < "3.12"
-anyio==3.7.0 ; python_version >= "3.8" and python_version < "3.12"
-appnope==0.1.3 ; python_version >= "3.8" and python_version < "3.12" and sys_platform == "darwin" or python_version >= "3.8" and python_version < "3.12" and platform_system == "Darwin"
-argon2-cffi-bindings==21.2.0 ; python_version >= "3.8" and python_version < "3.12"
-argon2-cffi==21.3.0 ; python_version >= "3.8" and python_version < "3.12"
-arrow==1.2.3 ; python_version >= "3.8" and python_version < "3.12"
-asttokens==2.2.1 ; python_version >= "3.8" and python_version < "3.12"
-atomicwrites==1.4.1 ; python_version >= "3.8" and python_version < "3.12" and sys_platform == "win32"
-attrs==23.1.0 ; python_version >= "3.8" and python_version < "3.12"
-babel==2.12.1 ; python_version >= "3.8" and python_version < "3.12"
-backcall==0.2.0 ; python_version >= "3.8" and python_version < "3.12"
-beautifulsoup4==4.12.2 ; python_version >= "3.8" and python_version < "3.12"
-bleach==6.0.0 ; python_version >= "3.8" and python_version < "3.12"
-certifi==2023.5.7 ; python_version >= "3.8" and python_version < "3.12"
-cffi==1.15.1 ; python_version >= "3.8" and python_version < "3.12"
-charset-normalizer==3.1.0 ; python_version >= "3.8" and python_version < "3.12"
-colorama==0.4.6 ; python_version >= "3.8" and python_version < "3.12" and platform_system == "Windows" or python_version >= "3.8" and python_version < "3.12" and sys_platform == "win32"
-comm==0.1.3 ; python_version >= "3.8" and python_version < "3.12"
+colorama==0.4.6 ; python_version >= "3.8" and python_version < "3.12" and platform_system == "Windows"
 contourpy==1.1.0 ; python_version >= "3.8" and python_version < "3.12"
-coverage[toml]==7.2.7 ; python_version >= "3.8" and python_version < "3.12"
 cycler==0.11.0 ; python_version >= "3.8" and python_version < "3.12"
-debugpy==1.6.7 ; python_version >= "3.8" and python_version < "3.12"
-decorator==5.1.1 ; python_version >= "3.8" and python_version < "3.12"
-defusedxml==0.7.1 ; python_version >= "3.8" and python_version < "3.12"
-docutils==0.17.1 ; python_version >= "3.8" and python_version < "3.12"
-exceptiongroup==1.1.1 ; python_version >= "3.8" and python_version < "3.11"
-executing==1.2.0 ; python_version >= "3.8" and python_version < "3.12"
-fastjsonschema==2.17.1 ; python_version >= "3.8" and python_version < "3.12"
-fonttools==4.40.0 ; python_version >= "3.8" and python_version < "3.12"
-fqdn==1.5.1 ; python_version >= "3.8" and python_version < "3.12"
-idna==3.4 ; python_version >= "3.8" and python_version < "3.12"
-imagesize==1.4.1 ; python_version >= "3.8" and python_version < "3.12"
-importlib-metadata==6.7.0 ; python_version >= "3.8" and python_version < "3.10"
-importlib-resources==5.12.0 ; python_version >= "3.8" and python_version < "3.10"
-iniconfig==2.0.0 ; python_version >= "3.8" and python_version < "3.12"
[…]
