+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "spikeometric"
-version = "1.0.1"
+version = "1.0.2"
 description = "Spikeometric is a Pytorch Geometric based framework for simulating Spiking Neural Networks using Linear Non-linear Cascade models"
 authors = ["Jakob Sønstebø <jakobls16@gmail.com>"]
 readme = "README.md"
@@ -11,7 +11,7 @@ packages = [{include = "spikeometric"}]
 
 [tool.poetry.dependencies]
 python = ">=3.8, <3.12"
-numpy = "^1.21.4"
+numpy = "<1.24"
 matplotlib = "^3.5.0"
 tqdm = "^4.62.3"
 scipy = "^1.7.3"
@@ -25,7 +25,12 @@ pytest-lazy-fixture = "^0.6.3"
 sphinx = "^4.3.2"
 sphinx-rtd-theme = "^1.1.1"
 jupyter = "^1.0.0"
-ipython = "^8.8.0"
+ipython = "^7.34.0"
+ipykernel = "^5.5.6"
+pandas = "^1.5.3"
+notebook = "^6.5.5"
+tornado = "6.3.2"
+decorator = "<5.0"
 
 [build-system]
 requires = ["poetry-core"]
