### 2023-09-05 bioAI-Oslo/Spikeometric#12 @JakobSonstebo
Hi! Thank you for bringing this to our attention. I will look into to it now.
