### 2023-09-05 8536263
Updated version numbers in requirements to match colabs preinstalled packages
