+++ requirements.txt
@@ -3,16 +3,16 @@ contourpy==1.1.0 ; python_version >= "3.8" and python_version < "3.12"
 cycler==0.11.0 ; python_version >= "3.8" and python_version < "3.12"
 fonttools==4.42.1 ; python_version >= "3.8" and python_version < "3.12"
 importlib-resources==6.0.1 ; python_version >= "3.8" and python_version < "3.10"
-kiwisolver==1.4.5 ; python_version >= "3.8" and python_version < "3.12"
-matplotlib==3.7.2 ; python_version >= "3.8" and python_version < "3.12"
+kiwisolver==1.4.4 ; python_version >= "3.8" and python_version < "3.12"
+matplotlib==3.7.1 ; python_version >= "3.8" and python_version < "3.12"
 networkx==3.1 ; python_version >= "3.8" and python_version < "3.12"
 numpy==1.23.5 ; python_version < "3.12" and python_version >= "3.8"
 packaging==23.1 ; python_version >= "3.8" and python_version < "3.12"
 pandas==1.5.3 ; python_version >= "3.8" and python_version < "3.12"
-pillow==10.0.0 ; python_version >= "3.8" and python_version < "3.12"
-pyparsing==3.0.9 ; python_version >= "3.8" and python_version < "3.12"
+pillow==9.4.0 ; python_version >= "3.8" and python_version < "3.12"
+pyparsing==3.1.1 ; python_version >= "3.8" and python_version < "3.12"
 python-dateutil==2.8.2 ; python_version >= "3.8" and python_version < "3.12"
-pytz==2023.3.post1 ; python_version >= "3.8" and python_version < "3.12"
+pytz==2023.3 ; python_version >= "3.8" and python_version < "3.12"
 scipy==1.10.1 ; python_version >= "3.8" and python_version < "3.12"
 seaborn==0.12.2 ; python_version >= "3.8" and python_version < "3.12"
 six==1.16.0 ; python_version >= "3.8" and python_version < "3.12"
