+++ pyproject.toml
@@ -12,11 +12,15 @@ packages = [{include = "spikeometric"}]
 [tool.poetry.dependencies]
 python = ">=3.8, <3.12"
 numpy = "<1.24"
-matplotlib = "^3.5.0"
+matplotlib = "^3.7.1"
 tqdm = "^4.62.3"
 scipy = "^1.7.3"
 seaborn = "^0.12.2"
 networkx = "^3.0"
+pyparsing = "^3.1.1"
+pillow = "9.4.0"
+kiwisolver = "1.4.4"
+pytz = "2023.3"
 
 [tool.poetry.dev-dependencies]
 pytest = "^6.2.5"
