### 2023-09-06 bioAI-Oslo/Spikeometric#14 @Saran-nns
Automated tests for device, stimulation and models failed
Hi team,

As part of [review](https://github.com/openjournals/joss-reviews/issues/5451#issuecomment-1682651871)

Error Traceback:

`Traceback (most recent call last):
  File "/content/Spikeometric/tests/test_stimulation.py", line 5, in <module>
    @pytest.mark.parametrize("stimulus", [pytest.lazy_fixture('regular_stimulus'), pytest.lazy_fixture('sin_stimulus'), pytest.lazy_fixture('poisson_stimulus')])
  File "/usr/local/lib/python3.10/dist-packages/pytest/__init__.py", line 171, in __getattr__
    raise AttributeError(f"module {__name__} has no attribute {name}")
AttributeError: module pytest has no attribute lazy_fixture`

Ref: https://colab.research.google.com/drive/1l8Y3So5uBXl4kA_6jvpkvyZOILVjdrHK#scrollTo=dXRS0vx6lJow

thanks
Saran
