### 2023-09-06 bioAI-Oslo/Spikeometric#15 @Saran-nns
Documentation not up-to-date
Hi Team,

Test run for an example in the documentation breaks.

Please check the error at https://colab.research.google.com/drive/1l8Y3So5uBXl4kA_6jvpkvyZOILVjdrHK#scrollTo=Ya0jZquvtSXZ&line=6&uniqifier=1

Also, for reference, it would be useful to mention the imports at each code blocks

thanks,
saran

https://github.com/openjournals/joss-reviews/issues/5451#issuecomment-1682651871
