### 2023-09-07 bioAI-Oslo/Spikeometric#14 @JakobSonstebo
Hi!

To run the tests in colab you need to install pytest-lazy-fixture and use pytest (run either "!pytest" to run all of them or "!pytest tests/test_file_you_want_to_run.py" to run one at a time). Also note that this needs to be done in the main directory for pytest to find the right test files. pytest-lazy-fixture is included in requirements_dev.txt but since that also includes jupyter related packages which colab doesn't like I recommend just pip installing it seperately.

Hope this helps,
Jakob
