### 2023-09-07 8458790
Corrected parameter names in RegularStimulus from interval and duration to period and stop in documentation
