### 2023-09-07 bioAI-Oslo/Spikeometric#15 @JakobSonstebo
Hi!

Thank you, you're right this was not updated last time a change was made to the parameter names. I've corrected interval to period and duration to stop in the introduction document now. Please let me know if you find any other inconsistencies between the documentation and the source, so I can correct them as quickly as possible.

Jakob
