### 2023-09-08 1539276
Added the batch size parameter to stimuli classes in docs and examples, included LoadedStimulus in implement_model tutorial and example and fixed typos
