+++ docs/tutorials/implement_model.rst
@@ -111,13 +111,13 @@ Stimulus filter
 ----------------
 We want the model to handle a multidimensional stimulus given to some of the neurons, 
 and must implement a :py:meth:`stimulus_filter` method to determine how this stimulus will be integrated into an input. 
-The function :py:func:`k` that we pass to the model will help us with this. 
-Note that we are assuming that all targeted neurons receive the same stimulus and that they have the same stimulus filter.
+The function :py:func:`k` that we pass to the model will decide how the different channels in the multidimensional stimulus are weighted.
+Note that we are assuming that all targeted neurons they have the same stimulus filter.
 
 .. code-block:: python
 
     def stimulus_filter(self, stimulus: torch.Tensor) -> torch.Tensor:
-        return torch.sum(stimulus*self.k(torch.arange(stimulus.shape[0])))
+        return torch.sum(stimulus*self.k(torch.arange(stimulus.shape[1])), dim=1)
 
 Input
 ------
@@ -133,7 +133,7 @@ Non-linearity
 --------------
 After we have collected the input to the neurons, we need to apply a non-linearity to it. In this case, 
 we use a rectified linear non-linearity. The rate is scaled by the time step to ensure that increasing the resolution
-of the simulation doesn't artificially increase the firing rate.
+of the simulation doesn't artificially change the firing rate.
 
 .. code-block:: python
 
@@ -196,7 +196,7 @@ Summary
             return W.flip(1)
 
         def stimulus_filter(self, stimulus: torch.Tensor) -> torch.Tensor:
-            return torch.sum(stimulus*self.k(torch.arange(stimulus.shape[0])))
+            return torch.sum(stimulus*self.k(torch.arange(stimulus.shape[1])), dim=1)
         
         def input(self, edge_index: torch.Tensor, W: torch.Tensor, state: torch.Tensor, t=-1) -> torch.Tensor:
             return self.synaptic_input(edge_index, W, state) + self.stimulus_input(t)
@@ -209,30 +209,39 @@ Summary
 
 Testing the model
 ------------------
-To test the model, we will simulate some networks with 50 neurons. We will use a random connectivity matrix  with normal weights and a random periodic stimulus.
+To test the model, we will simulate some networks with 50 neurons. We will use a random connectivity matrix with normal weights and a periodic stimulus
+with random amplitudes in five frequency bins. Since we are batching the simulations in groups of 2 networks,
+we will first generate a stimulus plan and make use of the :py:class:`LoadedStimulus` class to take care of the batching for us.
 
 .. code-block::
 
     n_neurons = 50
-    test_data = NormalGenerator(n_neurons, mean=0, std=0.1, glorot=True).generate(10)
-    loader = DataLoader(test_data, batch_size=5, shuffle=False)
+    test_data = NormalGenerator(n_neurons, mean=0, std=0.5, glorot=True).generate(10)
+    loader = DataLoader(test_data, batch_size=2, shuffle=False)
 
     def r(t):
-        return -100.*(t < 2) + -100*torch.exp(-2*(t-2))*(t >= 2)*(t<5)
+        return -1000.*(t < 2) + -1000*torch.exp(-2*(t-2))*(t >= 2)*(t<5)
 
     def w(t):
         return torch.exp(-t/2)
 
     def k(f):
-        return torch.exp(-f/5) # weight lower frequencies more
+        return torch.exp(-f)
 
-    def stimulus(t):
-        targets = torch.isin(torch.arange(50), torch.arange(50)[::10]).unsqueeze
-        return 0.1*torch.rand(5) * (t % 100 < 20)*targets
+
[…]
