### 2023-09-08 bioAI-Oslo/Spikeometric#14 @Saran-nns
thanks Jakob
