### 2023-09-08 #5451 @JakobSonstebo
Hi @Saran-nns, thank you for your suggestions. An example of how to use the LoadedStimulus class can be found in the simulate_with_stimulus.ipynb file in the examples directory.  I notice, however, that it doesn't show up in the overview over the available stimulus classes on the documentation page and will add it there now. For the batching of stimulus, the stimulus classes should be passed a batch_size parameter matching what is sent to the DataLoader class. I'll make sure to edit the examples and documentation to make sure this comes across more clearly.
