### 2023-09-11 47796b0
Added a benchmarking sections detailing memory usage and performance of the package
