### 2023-09-13 f62c636
Update poisson_glm_model.py
add information about the coupling filter to and changed tau to t' in the main equation to accomodate the parameter in the initialization
