### 2023-09-13 70fade4
Added missing DOI and fixed typo in line 62 of paper
