+++ paper.md
@@ -44,7 +44,7 @@ There exists a number of other simulation tools for simulating networks of neuro
 
 The `spikeometric` package is a framework for simulating spiking neural networks using linear non-linear cascade models in Python. It is built on the PyTorch Geometric package and uses its powerful graph neural network modules and efficient graph representation. It is designed to be flexible and easy to use, while also being competitive on speed. Moreover, it’s built in a way that accommodates usage or implementation of multiple different models sharing principles from the linear non-linear cascade family of models.
 
-The `torch` backend makes simulating large networks on a GPU easy, with the extra benefit of having a familiar use pattern, reducing the friction of picking up a new tool. The package relies heavily on `PyTorch Geometric` [@Fey/Lenssen/2019], with the networks being represented as `torch_geometric` `Data` objects and the models extending the `MessagePassing` base class. The `PyTorch Geometric` framework is a popular deep learning framework originally designed for Graph Neural Networks (GNNs), a class of neural networks for learning graph-related data [@DBLP:journals/corr/abs-2104-13478]. It is the perfect setting for simulating neural networks with tunable parameters, allowing us to us to formulate the model’s equations naturally in terms of vertices and edges, and giving us access to easy automatic tuning of parameters e.g. to match a certain firing rate, provided that the nonlinearity in the model is differentiable. The tuning functionality allows for fitting arbitrary parameters and can provide a starting point for implementing encoding models.
+The `torch` backend makes simulating large networks on a GPU easy, with the extra benefit of having a familiar use pattern, reducing the friction of picking up a new tool. The package relies heavily on `PyTorch Geometric` [@Fey/Lenssen/2019], with the networks being represented as `torch_geometric` `Data` objects and the models extending the `MessagePassing` base class. The `PyTorch Geometric` framework is a popular deep learning framework originally designed for Graph Neural Networks (GNNs), a class of neural networks for learning graph-related data [@DBLP:journals/corr/abs-2104-13478]. It is the perfect setting for simulating neural networks with tunable parameters, allowing us to formulate the model’s equations naturally in terms of vertices and edges, and giving us access to easy automatic tuning of parameters e.g. to match a certain firing rate, provided that the nonlinearity in the model is differentiable. The tuning functionality allows for fitting arbitrary parameters and can provide a starting point for implementing encoding models.
 
 In addition to the models, the package includes dataset classes that can generate random connectivity matrices from a distribution or load pre-constructed connectivity matrices into `torch_geometric`’s `Data` objects to be passed straight to the model. These objects hold a sparse representation of our connectivity matrices and can be batched together to form isolated subgraphs of a big graph, letting us simulate many networks simultaneously.
 
