+++ paper.bib
@@ -121,7 +121,8 @@ url={https://doi.org/10.1038/nature07140}
   year    = {2007},
   volume  = {2},
   pages   = {1430},
-  number  = {4}
+  number  = {4},
+  doi = {10.4249/scholarpedia.1430}
 }
 
 @book{carnevale_hines_2006,
@@ -209,5 +210,6 @@ abstract={Over the past decade there has been a growing interest in the developm
       year={2023},
       eprint={2109.12894},
       archivePrefix={arXiv},
-      primaryClass={cs.NE}
+      primaryClass={cs.NE},
+      doi = {10.48550/arXiv.2109.12894}
 }
