### 2023-09-13 #5451 @JakobSonstebo
Hi again @Saran-nns!

I've fixed the typo and added the missing DOIs. Seems like it looks good now. Thank you for taking the time to go through our work and helping us improve it. Greatly appreciated.

Jakob
