### 2023-09-17 79b0466
Added the recurrence strength parameter r in BernoulliGLM to documentation and examples
