### 2023-09-18 #5451 @clinssen
Hi, thanks for all the updates. This looks very good and I'm happy to give it my thumbs up! I hereby consider all of my earlier points resolved, and the checklist to be fully completed.

Just a few small points:
- the name of the simulator is "Brian2" and not "Brain2"
- in the paper, you write: "This gives it a different scope from existing frameworks." Different how? Please elaborate.
- typo on the readthedocs page: "memory conumption"
