### 2023-09-18 b39da56
Fixed typo in paper and removed potentially confusing sentence
