+++ paper.md
@@ -38,7 +38,7 @@ This software expands on a simulator developed for testing novel reconstruction 
 
 Linear non-linear cascade models are much used in computational neuroscience and come in many flavors. Typical examples are SRM, LNP, GLM [@Gerstner2008; @gerstner_kistler_naud_paninski_2014; @10.3389/fnsys.2016.00109]. What unites these models is that they all model the spike response of a network through the same cascade-like sequence of steps. At each time step, a neuron receives input from its environment and converts it to a firing rate by a nonlinear function. This firing rate parametrizes a probability distribution from which spikes are drawn. This contrasts hard-threshold-based models, in which spikes are emitted whenever a variable representing the membrane potential exceeds a certain value. Although these models share many of the same underlying principles, no unifying framework currently permits easy implementation and direct comparison. 
 
-There exists a number of other simulation tools for simulating networks of neurons, notably `NEST` [@Gewaltig:NEST], `Neuron` [@carnevale_hines_2006], `Brain2` [@10.7554/eLife.47314] and `GENESIS` [@Bower2013]. While all of these are primarily directed at solving systems where the dynamics are given by a system of differential equations, NEST includes a limited selection of pre-built stochastic point-process models similar to the ones found in `spikeometric`. However, this selection currently consists of only two models, which are not included in the GPU compatible version `NEST-GPU` [@10.3389/fninf.2022.883333; @10.3389/fncom.2021.627620]. There are also some `torch`-native frameworks like `BindsNET` [@10.3389/fninf.2018.00089] and `snnTorch` [@eshraghian2023training] but these focus on learning rather than simulation and are currently limited to LIF models. In general, `spikeometric` is more similar to the simulation tools in its focus, but has the automatic parameter tuning and `torch`-compatibility of the Spiking Neural Network libraries. This gives it a different scope from existing frameworks. 
+There exists a number of other simulation tools for simulating networks of neurons, notably `NEST` [@Gewaltig:NEST], `Neuron` [@carnevale_hines_2006], `Brian2` [@10.7554/eLife.47314] and `GENESIS` [@Bower2013]. While all of these are primarily directed at solving systems where the dynamics are given by a system of differential equations, NEST includes a limited selection of pre-built stochastic point-process models similar to the ones found in `spikeometric`. However, this selection currently consists of only two models, which are not included in the GPU compatible version `NEST-GPU` [@10.3389/fninf.2022.883333; @10.3389/fncom.2021.627620]. There are also some `torch`-native frameworks like `BindsNET` [@10.3389/fninf.2018.00089] and `snnTorch` [@eshraghian2023training] but these focus on learning rather than simulation and are currently limited to LIF models. In general, `spikeometric` is more similar to the simulation tools in its focus, but has the automatic parameter tuning and `torch`-compatibility of the Spiking Neural Network libraries.
 
 # Implementation
 
