+++ paper.bib
@@ -116,7 +116,7 @@ url={https://doi.org/10.1038/nature07140}
 
 @ARTICLE{Gewaltig:NEST,
   author  = {Marc-Oliver Gewaltig and Markus Diesmann},
-  title   = {NEST (NEural Simulation Tool)},
+  title   = {{NEST} ({NE}ural {S}imulation {T}ool)},
   journal = {Scholarpedia},
   year    = {2007},
   volume  = {2},
@@ -127,7 +127,7 @@ url={https://doi.org/10.1038/nature07140}
 
 @book{carnevale_hines_2006,
   place={Cambridge},
-  title={The NEURON Book},
+  title={The {NEURON} Book},
   DOI={10.1017/CBO9780511541612},
   publisher={Cambridge University Press},
   author={Carnevale, Nicholas T. and Hines, Michael L.},
@@ -170,7 +170,7 @@ url={https://doi.org/10.1038/nature07140}
 
 @ARTICLE{10.3389/fninf.2022.883333,
   author={Tiddia, Gianmarco and Golosio, Bruno and Albers, Jasper and Senk, Johanna and Simula, Francesco and Pronold, Jari and Fanti, Viviana and Pastorelli, Elena and Paolucci, Pier Stanislao and van Albada, Sacha J.},   
-  title={Fast Simulation of a Multi-Area Spiking Network Model of Macaque Cortex on an MPI-GPU Cluster},      
+  title={Fast Simulation of a Multi-Area Spiking Network Model of Macaque Cortex on an {MPI-GPU} Cluster},      
   journal={Frontiers in Neuroinformatics},      
   volume={16},           
   year={2022},      
@@ -182,7 +182,7 @@ url={https://doi.org/10.1038/nature07140}
 
 @ARTICLE{10.3389/fncom.2021.627620,
 author={Golosio, Bruno and Tiddia, Gianmarco and De Luca, Chiara and Pastorelli, Elena and Simula, Francesco and Paolucci, Pier Stanislao},   
-title={Fast Simulations of Highly-Connected Spiking Cortical Models Using GPUs},      
+title={Fast Simulations of Highly-Connected Spiking Cortical Models Using {GPU}s},      
 journal={Frontiers in Computational Neuroscience},
 volume={15},
 year={2021},      
@@ -194,7 +194,7 @@ abstract={Over the past decade there has been a growing interest in the developm
 
 @ARTICLE{10.3389/fninf.2018.00089,
 	author={Hazan, Hananel and Saunders, Daniel J. and Khan, Hassaan and Patel, Devdhar and Sanghavi, Darpan T. and Siegelmann, Hava T. and Kozma, Robert},   
-	TITLE={BindsNET: A Machine Learning-Oriented Spiking Neural Networks Library in Python},      
+	title={BindsNET: A Machine Learning-Oriented Spiking Neural Networks Library in {P}ython},      
 	journal={Frontiers in Neuroinformatics},      
 	volume={12},      
 	pages={89},     
