### 2023-09-18 #5451 @jbytecode
@lepmik - The PR (https://github.com/bioAI-Oslo/Spikeometric/pull/28) includes some minor changes on the bibliography file. Please review the PR. If you are agreed with the changes, please merge. thank you in advance.
