+++ paper.md
@@ -9,7 +9,7 @@ tags:
     - linear non-linear poisson models
 authors:
     - name: Jakob L. Sønstebø
-    orcid: 0009-0009-0584-9293
+      orcid: 0009-0009-0584-9293
       affiliation: 1
     - name: Herman Brunborg
       affiliation: 2
