### 2023-09-19 #5451 @jbytecode
@lepmik - Thank you for merging. Now the paper and the bibtex seem better.

- Please read and check the whole paper one more time. Please correct any issues if exist.
- Create a version tag in software repository and report here.
- Use this version tag to create a new Zenodo archive. Please report the DOI here.
- Please ensure that the author information including DOI's and the paper title match the zenodo archive.

Thank you in advance.
